requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
blis==1.3.0
build==1.2.2.post1
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpathlib==0.21.0
confection==0.1.5
coverage==7.8.0
cryptography==44.0.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.11
distro==1.9.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
langcodes==3.5.0
language_data==1.3.0
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
openai==1.79.0
packaging==25.0
pathspec==0.12.1
pd-anonymiser @ file:///Users/stuartpaton/github/pd-anonymiser
phonenumbers==8.13.55
pip-tools==7.4.1
platformdirs==4.3.8
pluggy==1.5.0
preshed==3.0.9
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyproject_hooks==1.2.0
pytest==8.3.5
pytest-cov==6.1.1
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
safetensors==0.5.3
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.2
sympy==1.14.0
thinc==8.3.6
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
